New __str__ method implemented on _ContextContainmentSignature. The purpose is purely for rich visualization of the score tree during debugging.
Example:

   t = Sequential(Staff([Voice(scale(2))]) * 2)
   t[0].invocation.name = 'staff1'
   t[1].invocation.name = 'staff2'
   t[0][0].invocation.name = 'voicefoo'
   t[1][0].invocation.name = 'voicefoo'
   diatonicize(t)
   Beam(t.leaves)

   {
           \context Staff = "staff1" {
                   \context Voice = "voicefoo" {
                           c'8 [
                           d'8
                   }
           }
           \context Staff = "staff2" {
                   \context Voice = "voicefoo" {
                           e'8
                           f'8 ]
                   }
           }
   }

  abjad> print t.leaves[2].parentage._containmentSignature

       root: Sequential-4393200 (4393200)
      score: 
      staff: Staff-staff2
      voice: Voice-voicefoo
       self: Note-5334832'''

Here we ask for the containment signature of the e'8. What returns if the component ID of self, followed by the component ID of the first voice in the parentage of self, followed by the first staff in the parentage of self, followed by the first score in the parentage of self (ie, None), followed by the root of the parentage of self.

Other changes included in this commit include making all attributes of _ContextContainmentSignature private, moving _Context._ID up to _Component._ID, and enriching _Component._ID to include classname and Python ID. Also, _Parentage._containment has been renamed to _Parentage._containmentSignature.

Also, the logic of _Parentage._containmentSignature has been improved to handle *implicit voices*. For example, when t = Staff(scale(4)) there is an explicit staff with four explicit notes. But there is no explicit voice. LilyPond creates a voice, however, at interpretation time and so now _Parentage._containmentSignature recognizes this situation and now derives _ContextContainmentSignature._voice from the explicit Staff. The intuitive idea here is that the explicit staff knows how to create an implicit voice, and, in fact, chooses to do so.

All navigator tests now pass. Containment signature comparison is now the best method available to determine whether two components 'are threadable' or 'have a path between them' or, perhaps better, 'belong to the same voice'.

One last detail. Signature comparison guarantees that two Abjad components are in the same voice. But signature comparison does not guarantee that there is an *uninterrupted* 'temporal pathway' between two Abjad components. That is, t1 and t2 may exhibit the same signature but be separate by a 'gap' of 40 measures. Contiguity testing should be implemented completely separately so that voice-resolution and contiguity testing can be mixed and matched independently for different applications.