Bug fix to _Parentage._threadSignature. Previously, _Parentage._threadSignature was using the concept of 'found_leaf' to find cases where leaves occur enclosed inan explicit staff but not in an explicit voice. This works fine for leaves. But this fails when measures or tuplets are enclosed in an explicit staff but not enclosed in an explicit voice.
So this commit extends the concept of 'found_leaf' to the concept of 'found_less_than_voice'. The new _is_less_than_voice(expr) helper committed in the last revision identifies all Abjad components like leaves, tuplets and measures that are somehow 'less than' a voice, which probably means that such components can be *contained in* a voice.

The newest test in parentage/test/test_parentage_thread_signature.py captures this case:

def test_parentage_thread_signature_10( ):
   '''Measure and leaves must carry same thread signature.'''

   t = Staff([DynamicMeasure(scale(2))] + run(2))
   diatonicize(t)

   r'''\new Staff {
         \time 1/4
         c'8
         d'8
      e'8
      f'8
   }'''

   assert t[0].parentage._threadSignature == t[-1].parentage._threadSignature
   assert t[0].parentage._threadSignature == t[0][0].parentage._threadSignature
   assert t[0][0].parentage._threadSignature == t[-1].parentage._threadSignature

That is all.