Major changes to the Spanner public interface.
Summary: Spanner no longer supports public __delitem__, __setitem__, insert( ) or remove( ). Nor does Spanner insert an optional index argument pop( ). All four of these methods could previously cause discontiguities in the component content of a spanner.

Note however that public insert( ) has now become private _insert( ) and is still available for helper functions that need to make behind-the-scene changes to spanner positioning during some other, more complex operation. Also, Spanner._severComponent(component) functions exactly as deprecated Spanner.remove(component) used to function. So Spanner._severComponent(component) can likewise be used by helper functions in the middle of some other operation.

Given all this, however, it still makes sense in the implementation of helper functions to stick to the public composer interface wherever possible.

Many test files changed to accommodate this functionality *removed* from the spanner public interface.

New spanner public interface:

   * Spanner.append(component)
   * Spanner.append_left(component) [new]
   * Spanner.copy( )
   * Spanner.clear( )
   * Spanner.extend(components)
   * Spanner.extend_left(components) [new]
   * Spanner.fracture(i, direction)
   * Spanner.fuse(spanner)
   * Spanner.index(component)
   * Spanner.pop( )
   * Spanner.pop_left( ) [new]
   * Spanner.trim(component)
   * Spanner.trim_left(component)

Test files pending for trim( ) and trim_left( ). The intention of the method is to remove everything in spanner from component rightwards (for trim( )) or from component leftwards (for trim_left( )), thus shortening spanner by potentially many components at once.

The notion of capture( ) and capture_left( ) methods that can work on a numeric count of components to capture (rather than on a reference list of components ot append or extend) probably won't work as well now that spanners can attach not just to leaves but to containers. The reason being that it is no longer completely clear which component(s) to the immediate right or left of the ends of spanner the spanner should capture: the next measure, or the tuplet beginning at the start of the next measure, or the first note in the tuplet at the start of the next measure? So no capture( ) methods added now in favor of reference only addition, at least for the time being.

All tests pass, though the code is a bit messy. This completes half the work of changing the spanner / container interaction patterns. The next several commits will extend container behavior to actively manage any hovering spanner context on the addition and removal of container contents.