The check(expr, runtime = 'composition') helper now takes an optional 'runtime' keyword argument. Valid values for the 'runtime' keyword include 'composition' and 'format', defaulting to 'composition'.
What this comes out to mean is that default use of check(expr) no longer check for empty containers. This is because Abjad allows empty containers at composition-time. Calling check(expr, runtime = 'format'), however, DOES check for empty containers. This is because LilyPond does NOT allow empty containers at format-time.

Here's the new doc string to the helper:

      Run every check in the 'check' module against expr.

      Set the 'runtime' keyword to 'composition' to run
      only those checks appropriate at composition-time.
      Abjad allow empty containers at composition-time.
      Empty container checks will NOT run.

      abjad> check(Voice([ ]), runtime = 'composition')
      True

      Set the 'runtime' keyword to 'format' to run the 
      complete set of checks appropriate at format-time.
      LilyPond does NOT allow empty containers at format-time.
      Empty container check WILL run.

      abjad> check(Voice([ ]), runtime = 'format')
      False

So it is now good to call check(expr) in ALMOST EVERY reg test.