Fairly major change to the behavior of _Component.detach( ).
Previous behavior: t.detach( ), for Abjad component t, previously caused t to detach from parentage and then to detach from spanners. This would always work for leaves. But in the case that t was a container somewhere in the middle of the score tree, this could leave the score tree in a weird state. Here's an example:

   t = Voice(run(2))
   t.insert(1, Sequential(run(2)))
   diatonicize(t)
   Beam(t.leaves)

   \new Voice {
      c'8 [
      {
         d'8
         e'8
      }
      f'8 ]
   }

   t[1].detach( )

At this point the sequential, t[1], has detached from voice and detached from beam. BUT t[1] continues to hold on to its children, d'8 e'8. And this is a problem because the beam continues to attach to c'8 e'8 e'8 f'8. Running check(t) returns False. Running report(t) shows 1 / 1 spanner discontiguous.

So, what's needed here are public methods that leave all score trees always in tact. (We might call these operations composer-safe or public-facing, or something). In the case of t, above, with sequential t[1], at least two composer-safe operations make good sense.

SLIP. The first is to 'slip' t[1] out of the score tree altogether, leaving d'8 e'8 to attach to voice directly. As of the previous commit, slipping does now work for all Abjad components. And slipping leaves all score trees always in tact.

(STRONGER) DETACH. The second operation that makes sense here is *stronger* form of detach. Specifically, an operation that detaches component from parentage, detaches component from spanners, AND ALSO detaches children of component from spanners. This strong form of detach is implemented as detach_subtree(component) in this commit. This commit also aliases _Component.detach( ) to this new detach_subtree( ).

What this means is that, as of this commit, the behavior of _Component.detach( ) has been 'strengthened' considerably to remove not just component from score tree but also the remove all children of component from score tree. Thus, t.detach( ) now leaves all score trees always in tact, for any Abjad component t.