Major change to Container.__setitem__(i, expr).
Container.__setitem__(i, expr) was previously set expr into self at the index of i, but without accounting either for spanners attaching directly to the current component at index i, or without accounting for spanners attaching to the children of the current component at index i. This means that the previous implementation of container __setitem__ could leave spanners in a weird state after application.

This new implementation of container __setitem__ leaves all score trees always in tact. Notionally, the way that this happens is that the container looks not just at the component sitting (shallowly) at index . Rather, the container looks at the *entire subtree* sitting at index i. Then the component asks "what dominant spanners exist that cover not just the component at index i but, indeed, *any of the children* of the component sitting at index i?" In other words, the container looks for and collects all spanners that dominate the subtree rooted at the component at index i. After finding all spanners that dominate the subtree rooted at the component at index i, the container then detaches the subtree rooted at the component at index i; this means that the component at index i, as well as all of the children of the component at index i, detach from all spanners. Following this the container inserts the new expression at index i. Lastly, and importantly, the container REAPPLIES the spanners that dominated index i to the new inhabitant of index i. This means that the new expression inherits ALL of the spanners that previously dominated index i. Another way of looking at this is that spanners that dominate index i prior to __setitem__ continue to dominate index i after __setitem__.

Example:

   t = Voice(Sequential(run(2)) * 2)
   diatonicize(t)
   Beam(t[:])
   Glissando(t[:])

   \new Voice {
      {
         c'8 [ \glissando
         d'8 \glissando
      }
      {
         e'8 \glissando
         f'8 ]
      }
   }

   t[1] = FixedDurationTuplet((2, 8), scale(3))

   \new Voice {
      {
         c'8 [ \glissando
         d'8 \glissando
      }
      \times 2/3 {
         c'8 \glissando
         d'8 \glissando
         e'8 ]
      }
   }

   assert check(t)

See container/test/test_container_setitem_integer.py for this example and five others. It is recommended to read all six examples in this test file extremely carefully to observe the important new behavior of this method.

Note carefully that the SLICE-VALUED version of Container.__setitem__( ) has not yet reimplemented! The slice-valued version of the method will reimplement soon in future commits.