New tools/imports package available. Use this package to import all functions from any package or module into any other package. The package contains three functions: _import_functions_in_package_to_namespace(package, namespace) _get_functions_in_modules(module_file) _get_topmost_module(module_file)
The getters are helpers, so they may me ignored. The function to work with is _import_functions_in_package_to_namespace(package, namespace).
It imports all the functions defined in the modules of the package given as a string path into the given namespace dictionary.
Typical use would be to place the function in the  __init__.py file of a package to load all functions in the modules of the package into the package's namespace.
   Example:
   A package structure like so:
      package.mod1.mod1_func1( )
      package.mod2.mod2_func1( )
      package.mod2.mod2_func2( )
      package.mod3.mod3_func1( )
   Ends up as
      package.mod1_func1( )
      package.mod2_func1( )
      package.mod2_func2( )
      package.mod3_func1( )

To do this, put the following lines in the __init__.py file of the package you want to display this behavior:
## start
from abjad.tools.imports.import_functions_in_package_to_namespace import _import_functions_in_package_to_namespace

_import_functions_in_package_to_namespace(__path__[0], globals( ))
## end

The new tools/imports package imports its own functions in this manner. 
The mathtools module will be restructured into a package to conform with the standard Abjad coding standard of having a function per module and will instroduce the usage of this tool. 

Note that the main Abjad __init__.py still does not know how to differenciate between packages whose functions should be automatically loaded into the Abjad namespace (the default behavior) and those that should not. This is why the imports package is currently inside the tools directory.