New container_insert_and_fracture(container, i, component) public helper. The helper copies the functionality of old Container.insert(i, component) exactly.
   Insert component into container at index i.
   Fracture spanners to the left of index i.
   Fracture spanners to the right of index i.
   Return Python list of fractured spanners.

   For nonfracturing insert, use Container __setitem__.

   Example:

   t = Voice(scale(4))
   Beam(t[:])
   container_insert_and_fracture(t, 1, Rest((1, 4)))

   \new Voice {
           c'8 [ ]
           r4
           d'8 [
           e'8
           f'8 ]
   }

The point of this commit is to externalize and preserve the old Container.insert( ) functionality so that the new Container.insert( ) implementation can wrap Container.__setitem__ exactly. So insert( ) will continue to be part of the public inteface of container. But insert( ) will soon switch implementation from a spanner-fracturing operation to a spanner-preserving operation.

See helpers/test/test_container_insert_and_fracture.py for seven examples total.