New _get_contained_spanners(components) private helper.
The helper asserts 'components' equal to a thread-contiguous list of Abjad components. The helper then iterates the components in 'components' and collects all spanners contained within any of the components in 'components', returning the result as an unordered set.

Example:

   t = Voice(Sequential(run(2)) * 2)
   diatonicize(t)
   beam = Beam(t[0][:])
   slur = Slur(t[1][:])
   trill = Trill(t.leaves)

   \new Voice {
           {
                   c'8 [ \startTrillSpan
                   d'8 ]
           }
           {
                   e'8 (
                   f'8 ) \stopTrillSpan
           }
   }

   spanners = _get_contained_spanners(t[0:1])
   assert len(spanners) == 2
   assert beam in spanners
   assert trill in spanners

See helpers/test/test_get_contained_spanners.py for this example.