Container.__setitem__(i, expr) with integer i now calls _withdraw_from_crossing_spanners( ) on 'expr' prior to setting 'expr' in container.
What this does is eliminate a potential spanner-mangling error in the way that integer-valued setitem has worked up to this point. Specifically, when stealing a component from one (spanned) donor container for setting in another, receipient container, it was, up until this commit, likely to cause a spanner-mangling problem. The reason is that the spanners attaching to the donor component would continue to attach to the donor component even after severing the *parentage* of the donor component and setting the donor component in the receipient container.

What the call to _withdraw_from_crossing_spanners(expr) does is make sure that the (spanned) donor component is free of crossing spanners before setting in the receipient container. Note also that _withdraw_from_crossing_spanners(expr) leaves any *COVERED* spanners contained in the donor component in tact. That is, the call free the donor component of any *CROSSING* spanners but leaves any *COVERED* spanners attached to the donor component.

This package of behaviors is easiest to see in the two new tests added in to container/test/test_container_setitem_integer.py and reproduced in their entirety here:

Container setitem with integer-value index example 8:

   Take spanned leaf from donor container 
   and insert into recipient container.
   Both donor and recipient check after set item.

   notes = scale(6)

   t = Voice(notes[:3])
   Beam(t[:])

   \new Voice {
      c'8 [
      d'8
      e'8 ]
   }

   u = Voice(notes[3:])
   Beam(u[:])
   \new Voice {
      f'8 [
      g'8
      a'8 ]
   }

   t[1] = u[1]

   Modified t

   \new Voice {
      c'8 [
      g'8
      e'8 ]
   }

   assert check(t)

   Modified u:

   \new Voice {
      f'8 [
      a'8 ]
   }

   assert check(u)

Note carefully that both the donor and recipient containers check fully after setitem.


Container setitem with integer-value index example 9:

   Take down-spanned container with completely covered spanner 
   from donor container and insert into recipient container.
   Both donor and recipient check after set item.

   notes = scale(7)

   t = Voice(notes[:3])
   Beam(t[:])

   \new Voice {
      c'8 [
      d'8
      e'8 ]
   }

   u = Voice(notes[3:])
   Sequential(u[1:3])
   Glissando(u.leaves)
   Slur(u[1].leaves)

   \new Voice {
      f'8 \glissando
      {
         g'8 \glissando (
         a'8 \glissando )
      }
      b'8
   }

   t[1] = u[1]

   "Voice t is now ..."

   \new Voice {
      c'8 [
      {
         g'8 (
         a'8 )
      }
      e'8 ]
   }

   assert check(t)

   "Voice u is now ..."

   \new Voice {
      f'8 \glissando
      b'8
   }

   assert check(u)

Note carefully again that both t and u check fully after setitem.

What's going on here is a relatively sophisticated copy / paste behavior that winds up being important where 'stealing' a spanned component from a donor container and setting the component in a recipient container.