Added new _add(spanner) method to _ComponentSpannerAggregator.
Since _ComponentSpannerAggregator is basicaly a wrapper around the Python unordered set built-in, it makes sense for _ComponentSpannerAggregator to have an _add( ) method that corresponds to the built-in set.add( ) method. 

Up until this release, _ComponentSpannerAggregator had implemented only an _update(spanners) method which, like list.extend( ), demands an iterable of zero or more spanners. The new _add(spanner) method demands a single spanner.

Tests pending.