Major update to the way slice-valued container setitem works.
Slice-valued calls to setitem, such as t[i:j] = [components] previouslys made sure to detach every component in 'components' from any parent that component may have had. This is good. But such calls did NOT detach the components in 'components' from any _appropriate_ spanners attaching to the components in 'components' or attaching to the children of the components in 'components'. As it turns out, it is important to detach the _appropriate_ spanners attaching to the components in 'components' or attaching to the children of the components in 'components'.

So which spanners attaching to either the components in 'components' or to the children of the components in 'components' turn out to be _appropriate_ to detach as part of a slice-valued call to setitem, such as t[i:j] = [components]?

The answer is that those spanners which count as *CROSSING* spanners of the components in 'components' should be removed from the components in 'components'. Note further, however, that it is equally important to *preserve* (rather than remove) any spanners which count as *COVERED* spanners with respect to the components in 'components'.

The present commit implements these behaviors. That is, components to be set in a recipient container withdraw from crossing spanners prior to setting in the recipient container.

The consequences of this new behavior affect container append( ), extend( ) and insert( ), enabling all three methods to take components out of, and then set back into, the one and same container. That is, it is now possible for DONOR and RECIPIENT containers to be THE SAME. Some examples help show the new possibilities.

Example of container append with donor and recipient container being the same:

   t = Voice(scale(4))
   Beam(t[:])

   \new Voice {
      c'8 [
      d'8
      e'8
      f'8 ]
   }

   t.append(t[1])

   \new Voice {
      c'8 [
      e'8
      f'8 ]
      d'8
   }

   assert check(t)

Notice that spanners are not mangled and that t checks fully because the donor component withdraws from all crossing spanners prior to being set in the recipient container.


Example container extend with separate donor and recipient containers and with sophisticated spanner-handling:

   t = Voice(scale(2))
   Beam(t[:])

   \new Voice {
      c'8 [
      d'8 ]
   }

   u = Voice(scale(4))
   Beam(u[:])
   Slur(u[-2:])

   \new Voice {
      c'8 [
      d'8
      e'8 (
      f'8 ] )
   }

   t.extend(u[-2:])

   "Container t is now ..."

   \new Voice {
      c'8 [
      d'8 ]
      e'8 (
      f'8 )
   }

   assert check(t)
   assert t.format == "\\new Voice {\n\tc'8 [\n\td'8 ]\n\te'8 (\n\tf'8 )\n}"

   "Container u is now ..."

   \new Voice {
      c'8 [
      d'8 ]
   }

   check(u)

Notice that exactly the donor components withdraw from exactly the one crossing spanner and yet also preserve the one covered spanner. The result is that no spanners anywhere are mangled and that both the donor and recipient containers check fully after, respectively, surrendering and incorporating components.

New tests in the following files:

   * container/test/test_container_append.py
   * container/test/test_container_extend.py
   * container/test/test_container_setitem_slice.py

It is recommended to read all three test files in their entirety to understand the new behavior.