New _SpannerDurationInterface.
Changes to the public interface: previously, p.duration, for spanner p, referenced a read-only attributed that returned the sum of the *prolated* durations of each of the components in spanner. Now p.duration returns a read-only reference to a new _SpannerDurationInterface that each spanner gets at initialization-time. Use p.duration.prolated to return what used to be given by p.duration. Also, use p.duration.written to return the sum of the *written* durations of each of the components in spanner. This was previously p.written. The spanner namespace is now cleaned up somewhat. And this represents the first compartmentalizing of spanner attributes.

Example:

   t = Voice([RigidMeasure((2, 12), scale(2)), RigidMeasure((2, 8), scale(2))])
   beam = Beam(t.leaves)
   crescendo = Crescendo(t[0][:])
   decrescendo = Decrescendo(t[1][:])

   assert beam.duration.prolated == Rational(5, 12)
   assert crescendo.duration.prolated == Rational(2, 12)
   assert decrescendo.duration.prolated == Rational(2, 8)

   assert beam.duration.written == Rational(4, 8)
   assert crescendo.duration.written == Rational(2, 8)
   assert decrescendo.duration.written == Rational(2, 8)

Several test files changed.