Implemented _NoteHeadInterface on _Component. No changes to the public interface.
Summary: the Note class has had a Note.notehead attribute equal to a NoteHead instance for years. Three or four months ago, a new _NoteHeadInterface class implemented. But adding _Component.notehead equal to a _NoteHeadInterface instance caused all sorts of conflicts with existing the existing Note.notehead equal to a NoteHead instace. (Note carefully the difference between the longstanding NoteHead class and the newer _NoteHeadInterface class.) The workaround for this at the time was to implement Container.notehead as a _NoteHeadInterface, leaving Note.notehead equal to a NoteHead instance. This was suboptimal because the _NoteHeadInterface should have implemented at the top-level _Component, with the Note class still free to override whatever it needed to after _Component initialization.

The current commit remedies this, removes Container.notehead and implements top-level _Component.notehead instead. The key to making this work was causing the longstanding NoteHead class to inherit directly from the newer _NoteHeadInterface class.

Note that a bunch of old code had to be touched for this feature: note, rest, chord and skip code, together with their initializers. The changes were mostly small, and mostly represent improvements that had needed to take place for a very long time.

All tests pass.

One last note, however: it's becoming increasingly obvious that the casting code implemented in the different leaf initializers with calls to _transfer_all_attributes( ) is just too kludgy. The current casting code does in fact work, but that seems like something of a small miracle. The pattern impelemented there of inspecting object dictionaries and moving things over (but only moving SOME things over ... with an unstable enumeration of protected attributes needing to be maintained by _transer_all_attributes( )) is just begging for trouble.

I think that, at some point in future, we should just eliminate the casting code and _transfer_all_attributes( ) entirely. In place of that we should implement a way for different classes of leaves to 'clone' each other from very minimal information (such as possibly written duration alone), and then use the newer, cleaner implementation of bequeath( ), or other such methods and helpers, to handle reinsertion to the score after cloning.