Generalized the concept of the crossing spanner:
   A spanner P crosses a list of thread-contiguous components C
   when P and C share at least one component and when it is the
   case that NOT ALL of the components in P are also in C.
   In other words, there is some intersection -- but not total
   intersection -- between the components of P and C.

The previous definition of crossing spanners relied on the idea of a spanner crossing either the left time-boundary, or the right time-boundary, or both time-boundaries of a thread-contiguous list of components.

This new definition of crossing spanners makes no recourse to the concept of time-boundaries and uses set intersection instead.

I would worry that this new, intersection-based implementation of get_crossing_spanners(components) would be too slow when the length of 'components' is large. However, I think the built-in Python set type is optimized for these sorts of things, even with relatively large numbers of items in the set.

Example:

   t = Voice(RigidMeasure((2, 8), run(2)) * 3)
   diatonicize(t)
   beam = Beam(t[1:2] + t[2][0:1])

   r'''\new Voice {
         \time 2/8
         c'8
         d'8
         \time 2/8
         e'8 [
         f'8
         \time 2/8
         g'8 ]
         a'8
   }'''

   spanners = get_crossing_spanners(t.leaves)

   assert len(spanners) == 1
   assert beam in spanners

This new example now appears in helpers/test/test_get_crossing_spanners.py.