Spanners now raise ContiguityError at initialization-time.
Public interface stays the same. But trying to apply a spanner to components that are NOT thread-contiguous will now raise a ContiguityError.

Example:

   t = Staff(Voice(scale(2)) * 2)
   assert py.test.raises(ContiguityError, 'Beam(t.leaves)')

The reason for the ContiguityError here is that two separate (anonymous) voices grant their leaves separate thread signatures.

This is a fairly important initialization-time assertion to make and it should have been in a model for a while. There is the chance that existing piece code will break, but it seems relatively unlikely.

Did have to fix a handful of tests, though, mostly of the sort that (unthinkingly) spanned an entire staff at once.