Simplified f( ) to print ONLY the 'format' attribute of Abjad expressions.
This actually restricts the functionality of the f( ) helper somewhat. Previously the f( ) helper would iterate a Python list of Abjad expressions and print the format of each Abjad expression in list. Calling f( ) on a Python list of any sort now raises an AttributeError:

   abjad> notes = scale(4)
   abjad> f(notes)
   Traceback (most recent call last):
     File "<stdin>", line 1, in <module>
     File "/Users/trevorbaca/Documents/abjad/trunk/abjad/helpers/f.py", line 14, in f
       print expr.format
   AttributeError: 'list' object has no attribute 'format'

The advantage to the newer, restricted behavior of f( ) is that Abjad containers like measures, tuplets, voices and so that are in some way poorly formed will now raise an explicit exception at format-time. Previously, poorly formed Abjad containers would raise an exception that f( ) would swallow, and then print out the format of each of their internal components. Now the container is either formatted correctly or it isn't, which gives cleaner results at format-time.

The other way to look at this is that, if you want to see a list of Abjad components, you need to put them in an Abjad container.