All Abjad components now implement a public splice( ) method.
Example:

   t = Voice(scale(3))
   Beam(t[:])   
   t[-1].splice(scale(3))

   r'''\new Voice {
      c'8 [
      d'8
      e'8
      c'8
      d'8
      e'8 ]
   }'''

The method accepts a Python list of Abjad components and then inserts the components immediately after t, inserting the components in the parent of t, if any, and extending spanners that dominate t to attach directly to components.

There has been a splice_after(component, more_components) helper for a few weeks now. But a somewhat surprising discovery is that this functionality should actually be implemented as a method bound to all Abjad components rather than as an independent helper function. The reason for this is that no generalization of the functionality to splice after, for example, a list of many components, makes sense. It always only ever makes sense to splice after one, and only one, individual component. So it therefore makes best sense to implement splicing functionality as a method bound to all Abjad components.

This is an intermediate step in deprecating splice_after( ) in favor of _Component.splice( ). A final step will include the implementation of _Component.splice_left( ) or _Component.splice_before( ) as well.