All Abjad components now feature a splice_left(components) bound method.
The method assert that 'components' is a Python list of zero or more Abjad components and then inserts the components in 'components' immediately before t in the parent of t. The helper also extends all that spanners that dominate t (or any of the children of t) *leftwards* to encompass all components in 'components'.

Example:

   t = Voice(Container(run(2)) * 2)
   diatonicize(t)
   Beam(t.leaves)
   result = t[1].splice_left([Note(2.5, (1, 8))])

   r'''\new Voice {
           {
                   c'8 [
                   d'8
           }
           dqs'8
           {
                   e'8
                   f'8 ]
           }
   }'''

New test files:

   * leaf/test/test_leaf_splice_left.py
   * container/test/test_container_splice_left.py

See the two new test files for a total of four examples, including this one.

This commit completes the splicing interface for all Abjad components.