This commit modifies the _ThreadInterface.signature property to to mimic LilyPond's  behavior of treating parallel leaves outside an explicit Context as belonging to different threads. Examples:
<< c8 d8 e8 >>

\new Staff << c8 d8 e8 >>

Thread's become ambiguous when parallel structures appear in the absence of explicit Context and more specifically in the absence of explicit Voices. The new _ThreadInterface.signature property tries to resolve this by keeping track of any Voices found in the tree structure and by checking if the root of the caller is a parallel container. 

A couple of tests in test_assess_components_none_thread.py and test_navigator_is_threadable.py were modified to conform to this behavior.

Note that this is just a "patch" commit that does not fully resolve the Abjad threading model. A much better solution will be implemented in future commits where parallel containers will be *required* to contain only Contexts, not Leaves or other Containers. This conclusion was reached during the development of the present commit. In this future model, we will be able to resolve thread signatures unambiguously and without the need to consider exceptional cases.