Complex commit. There's a collection of related helpers from about a year ago. These are:
   * excise( )
   * lcopy( )
   * scopy( )
   * tcopy( )

There's quite a bit going on in each of these four functions, including dependencies on many other packages and modules.

So how to migrate these four modules from the helpers/ directory to a reasonable tools/ package?

Semantically, lcopy( ) is really a way of copying a certain number of contiguous leaves *including the parentage of the target leaves*. So lcopy( ) is a copy operation, which suggests that lcopy( ) might in fact be a type clone, suitable for inclusion in the tools/clone package. This is probably the case and, ideally, lcopy( ) should probably becomes something like clone.by_count_with_parentage( ). Likewise, scopy( ) is a way of copying a certain amount of 'content' -- that is, either contents of a container or some fractions duration of an individual leaf -- *including the parentage of the target content*. So scopy( ) is also a copy operation and, ideally, should probably be named clone.by_duration_with_parentage( ).

Early versions of this commit in fact tried to move lcopy( ) and scopy( ) into the tools/clone package. But this causes some really nasty circular dependencies. Thinking about it, it seems that there is no real reason for clone functions that include parentage to have so many external dependencies. But the current implementations of lcopy( ) and scopy( ) are loaded with external dependencies and therefore must be put somewhere other than tools/clone, at least for now.

This commit therefore introduces a tools/clonewp package. The 'wp' in the package name can be taken to stand for 'with_parentage'. This comit then makes the following changes:

   helpers/lcopy.py ==> tools/clonewp/by_count_with_parentage.py
   helpers/scopy.py ==> tools/clonewp/by_duration_with_parentage.py

At some point the modules in the tools/clonewp should be reimplmeented following the newer, cleaner patterns with correspondingly fewer external package dependencies; at that point it will then be possible to migrate such modules to the regular tools/clone package.