New doc_sphinx folder added to documentation. This folder will house the new Sphinx documentation. The folder is already Sphinx-configured.
To compile the documentation in HTML type $ make html inside the root directory of the sphinx documentation (i.e. in doc_sphinx). This will crate a _build folder with the compiled docs. This _build folder should not be commited to the repository. 

All the documentation proper, except for the main index, lives inside the chapters sub-folder. All document files end with .rst (for reStructured Text). 

Sphinx version 0.6.1 required to compile docs.