Fixed extremely tricky bug in spannertools.get_dominant( ).
Previous implementation of spannertools.get_dominant(components) was based on the start offset and stop offset of the components in 'components'. This breaks when some or all of the components in 'components' are empty containers. Empty containers have no duration. This means that empty containers can render the the start offset and stop offset of components in 'components' meaningless.

This new implementation of spannertools.get_dominant(components) looks at the contemporaneous start contents of components[0] and at the contemporaneous stop contents of components[-1] and asks what spanners include one component in contemporaneous start contents of components[0] and also include one component in contemporaneous stop contents of components[-1]. Those spanners, and only those spanners, which meet these two criteria are returned as the spanners that dominante the components in 'components'.

Example of spanners dominating empty containers:

   t = Voice(Container([ ]) * 3)
   beam = Beam(t[:2])
   glissando = Glissando(t[1:])
   trill = Trill(t.leaves)

   r'''\new Voice {
           {
           }
           {
           }
           {
           }
   }'''

   receipt = spannertools.get_dominant(t[:1])

   "Only beam dominates first container."

   assert len(receipt) == 1
   assert (beam, 0) in receipt

New test added.

Also fixed a subtle bug in fuse.measures_by_reference( ) that was based on precisely this previous, time-based implementation of spannertools.get_dominant( ).