Addition to tools/fuse package. New fuse.tuplets_by_reference(tuplets) function. Contents of package now stand at:
   fuse.containers_by_reference( )
   fuse.contents_by_count( )
   fuse.leaves_by_reference( )
   fuse.measures_by_count_cyclic( )
   fuse.measures_by_reference( )
   fuse.tuplets_by_reference( )

The new fuse.tuplets_by_reference( ) implements in much the same way as fuse.measures_by_reference( ). The function works on both in-score and outside-of-score tuplets. The function also works on both fixed-duration and fixed-multiplier tuplets.

Example:

   t1 = FixedDurationTuplet((2, 8), construct.scale(3))
   Beam(t1[:])
   t2 = FixedDurationTuplet((2, 16), construct.scale(3, Rational(1, 16)))
   Slur(t2[:])
   t = Voice([t1, t2])

   r'''\new Voice {
        \times 2/3 {
                c'8 [
                d'8
                e'8 ]
        }
        \times 2/3 {
                c'16 (
                d'16
                e'16 )
        }
   }'''

   fuse.tuplets_by_reference(t[:])

   r'''\new Voice {
        \times 2/3 {
                c'8 [
                d'8
                e'8 ]
                c'16 (
                d'16
                e'16 )
        }
   }'''

Tuplets to fuse must carry the same multiplier. Tuplets to fuse must also be the same type. The function raises the new TupletFuseError when either of these conditions is False.

See the test file for four additional examples.